requirements.txt
attrs==23.1.0
rich==13.5.2
GvGen==1.0
six==1.16.0
